feat: Update projects data in constants/index.js
The projects data in the constants/index.js file has been updated to include a new project called "TikGuard - 2024 TikTok TechJam". This change aligns with recent user commits and repository commits that updated the navigation bar and added the WorkExperience component. The purpose of this change is to showcase the user's education journey by including the new project in the projects section of the website.